backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
sqlalchemy
pydantic
pandas
numpy
scikit-learn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pymongo
